[util] Move output options into output-options.hh
