+++ util/hb-shape.cc
@@ -27,6 +27,7 @@
 
 #include "shape-consumer.hh"
 #include "shape-format.hh"
+#include "output-options.hh"
 #include "main-font-text.hh"
 
 const unsigned DEFAULT_FONT_SIZE = FONT_SIZE_UPEM;
