+++ util/view-cairo.hh
@@ -28,6 +28,7 @@
 #define VIEW_CAIRO_HH
 
 #include "view-options.hh"
+#include "output-options.hh"
 #include "helper-cairo.hh"
 
 struct view_cairo_t : view_options_t, output_options_t
